+++ src/theme/components/pagination.ts
@@ -9,6 +9,11 @@ const Pagination = {
       lineHeight: "1.15",
       textDecoration: "none",
     },
+    li: {
+      _first: {
+        marginTop: "xxs",
+      },
+    },
   },
 };